detail page api testing with docs backend api
