+++ .env
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_API_ENDPOINT=http://localhost:8080
