+++ next.config.js
@@ -1,9 +1,12 @@
 /** @type {import('next').NextConfig} */
-const path = require('path')
+const path = require('path');
 const nextConfig = {
-      sassOptions: {
-        includePaths: [path.join(__dirname, 'styles')],
-      },
+  sassOptions: {
+    includePaths: [path.join(__dirname, 'styles')],
+  },
+  images: {
+    domains: ['recipe1.ezmember.co.kr'],
+  },
 };
 
 module.exports = nextConfig;
