--- src/app/docs/css/page.module.scss
@@ -1,249 +1,0 @@
-.main {
-  background-color: white;
-}
-
-.docs__title {
-  background-color: white;
-  padding: 1vh 2vw;
-  border: 1px solid gray;
-  border-radius: 10px;
-  box-shadow: 2px 2px 2px gray;
-  margin: 15px 0;
-  p {
-    padding: 2vh 0;
-    font-size: 2vw;
-  }
-  img {
-    border-radius: 10px;
-    width: 100%;
-    height: 70vh;
-  }
-  div {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding-top: 0 2vh;
-    span {
-      display: flex;
-      align-items: center;
-      img {
-        background-color: gray;
-        width: 35px;
-        height: 35px;
-        border-radius: 17.5px;
-      }
-      p {
-        margin: 0 1vw;
-        font-size: 1.3vw;
-      }
-    }
-    div {
-      button {
-        background-color: #d2de32;
-        font-size: 1vw;
-        margin-right: 1vw;
-        border-radius: 15px;
-        width: 8vw;
-        height: 4vh;
-      }
-      .like {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        background-color: rgb(185, 186, 187);
-        border-radius: 5px;
-        width: 100px;
-        height: 40px;
-        margin-right: 0;
-        p {
-          margin-left: 1vw;
-          font-size: 1vw;
-        }
-      }
-    }
-  }
-}
-
-.docs__ingredient {
-  background-color: white;
-  border: 1px solid gray;
-  border-radius: 10px;
-  box-shadow: 2px 2px 2px gray;
-  margin: 15px 0;
-  padding: 2vh 2vw;
-  text-align: right;
-  h1{
-    text-align: left;
-    span{
-      padding: 0 10px;
-    }
-  }
-  .mycart__button {
-    border-radius: 30px;
-    padding: 10px 20px;
-    background-color: #d2de32;
-  }
-  .docs__ingredient__list {
-    display: flex;
-    overflow-x:auto;
-    background-color: white;
-    border-radius: 10px;
-    border: 1px solid #E4E6EA;
-    box-shadow: 2px 2px 2px gray;
-    margin: 15px 0;
-    div {
-      margin: 30px 0;
-      padding-left: 2vw;
-      width: 100%;
-      h2{
-        text-align: left;
-      }
-      li {
-        display: flex;
-        flex-grow: 1;
-        width: 43vw;
-        margin: 3px 0;
-        ul {
-          display: flex;
-          justify-content: space-between;
-          align-items: center;
-          width: 100%;
-          font-size: 1vw;
-          p {
-            font-size: 1vw;
-          }
-          button {
-            padding: 5px 10px;
-            border-radius: 10px;
-            background-color: #d2de32;
-          }
-        }
-      }
-    }
-    div:last-child{
-      padding-right: 2vw;
-    }
-  }
-  .docs__ingredient__list::-webkit-scrollbar {
-    width: 12px;
-  }
-  
-  .docs__ingredient__list::-webkit-scrollbar-thumb {
-    background-color: #d2de32; 
-    border-radius: 12px;
-  }
-  
-  .docs__ingredient__list::-webkit-scrollbar-track {
-    background-color: #CFCFCF;
-    border-radius: 12px;
-  }
-}
-
-.docs__ricipe {
-  background-color: white;
-  padding: 1vh 2vw;
-  border: 1px solid gray;
-  border-radius: 10px;
-  box-shadow: 2px 2px 2px gray;
-  margin: 15px 0;
-  h1{
-    span{
-      padding: 0 10px;
-    }
-  }
-  .docs__ricipe__list{
-    li{
-      display: flex;
-      margin:15px 0;
-      border-radius: 15px;
-      border: 1px solid #E4E6EA;
-      box-shadow: 2px 2px 2px gray;
-      ul{
-        display: flex;
-        justify-content: space-between;
-        flex-grow: 1;
-        flex-direction: row;
-        div{
-          width: 50px;
-          height: 50px;
-          background-color: #d2de32;
-          border-radius: 50px;
-          margin: 10px;
-          p{
-            text-align: center;
-            line-height: 50px;
-          }
-        }
-        .docs__ricipe__comment{
-          width: 65%;
-          text-align: left;
-          margin: 10px;
-        }
-        img{
-          width: 20%;
-          height: 90%;
-          border-radius: 15px;
-          margin: 10px 15px;
-        }
-      }
-    }
-  }
-}
-
-.docs__comment {
-  background-color: white;
-  padding: 1vh 2vw;
-  border: 1px solid gray;
-  border-radius: 10px;
-  box-shadow: 2px 2px 2px gray;
-  h1{
-    span{
-      padding: 0 10px;
-    }
-  }
-  .docs__comment__line{
-    border: 1px solid #CFCFCF;
-  }
-  div{
-    display: flex;
-    width: 100%;
-    border-radius: 15px;
-    margin: 10px 0;
-    background-color: #E4E6EA;
-    img{
-      width: 50px;
-      height: 50px;
-      margin: 15px;
-      border-radius: 25px;
-      background-color: #CFCFCF;
-    }
-    div{
-      display: flex;
-      flex-direction: column;
-      margin: 0;
-      margin-right: 15px;
-      div{
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-        margin-top: 15px;
-        margin-right: 0;
-        p{
-          margin-bottom: 3px;
-        }
-      }
-      span{
-        flex-grow: 1;
-        border-radius: 15px;
-        padding: 5px 10px;
-        background-color: white;
-        margin-bottom: 10px;
-      }
-    }
-  }
-  
-  .docs__comment__comment{
-    width: 95%;
-    margin-left: auto;
-  }
-}
