+++ src/app/mypage/css/page.module.scss
@@ -19,6 +19,11 @@
       button {
         padding: 5px;
         font-size: 25px;
+        .mypage__bar__icon {
+          margin-right: 3px;
+          width: 20px;
+          height: 20px;
+        }
       }
       :hover {
         background-color: white;
@@ -31,6 +36,10 @@
       button {
         padding: 5px;
         font-size: 25px;
+        .mypage__bar__icon {
+          width: 20px;
+          height: 20px;
+        }
       }
       :hover {
         background-color: white;
@@ -49,8 +58,4 @@
       }
     }
   }
-}
-
-.mypage__bar__icon {
-  margin-right: 3px;
 }
