+++ src/app/mypage/page.tsx
@@ -4,7 +4,6 @@ import styles from './css/page.module.css'
 import { faBook, faGear, faHeart, faRightFromBracket, faUser, faUsers } from '@fortawesome/free-solid-svg-icons'
 import MyRecipe from '../components/myrecipe/page'
 import MyFollow from '../components/myfollow/page'
-import MyProfile from '../components/myprofile/page'
 import MyLike from '../components/mylike/page'
 import { useState } from 'react'
 import MyInfo from '../components/myinfo/page'
